Modularize a bit
