+++ server/config.py
@@ -1,0 +1,3 @@
+class Config(object):
+    SQLALCHEMY_DATABASE_URI = 'mysql://cody:cody@localhost/cody'
+    DEBUG = True
